Organized exports for diff tool
